Ticket 48894 - Issue with high number of entry state objects.
Bug Description:  With high numbers of entry state objects, we saw performance
degredation with our sorting algorithm in valueset.

Fix Description:  This contribution was originally by Mohammad N. converted the
valueset to quicksort. I have changed this to the Hoare quicksort which uses
slightly less operations than the lomuto sort that was provided.

Additionally, this fixes valueset to size_t, which is correct for arrays, fixes
some of our asserts, cleans up braces in code, fixes a misleading indent.

https://fedorahosted.org/389/ticket/48894

Author: Mohammad N., wibrown

Review by: lkrispen (Thanks!)

+++ ldap/servers/slapd/slap.h
@@ -564,9 +564,9 @@ struct slapi_value
 #define VALUE_SORT_THRESHOLD 10
 struct slapi_value_set
 {
-	int num; /* The number of values in the array */
-	int max; /* The number of slots in the array */
-	int *sorted; /* sorted array of indices, if NULL va is not sorted */
+	size_t num; /* The number of values in the array */
+	size_t max; /* The number of slots in the array */
+	size_t *sorted; /* sorted array of indices, if NULL va is not sorted */
 	struct slapi_value **va;
 };
